pyproject.toml
[project]
name = "spreadcalculator"
version = "1.0.0"
description = ""
authors = [
    {name = "Beliy",email = "beliypozhizni@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "redis (>=7.3.0,<8.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "easy-logging @ git+https://github.com/Beliypozhizni/EasyLogging.git@v.1.0.0",
    "spreads-detector-schemas @ git+https://github.com/Beliypozhizni/SpreadsDetectorSchemas.git"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
